+++ metrics_test.go
@@ -1,3 +1,20 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
 package apm_test
 
 import (
@@ -78,14 +95,6 @@ func TestTracerMetricsBuiltin(t *testing.T) {
 		"system.process.cpu.total.norm.pct",
 		"system.process.memory.size",
 		"system.process.memory.rss.bytes",
-
-		"agent.send_errors",
-		"agent.spans.sent",
-		"agent.spans.dropped",
-		"agent.transactions.sent",
-		"agent.transactions.dropped",
-		"agent.errors.sent",
-		"agent.errors.dropped",
 	}
 	sort.Strings(expected)
 	for name := range builtinMetrics.Samples {