+++ module/apmot/harness_test.go
@@ -1,3 +1,20 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
 package apmot
 
 import (
@@ -7,6 +24,7 @@ import (
 
 	opentracing "github.com/opentracing/opentracing-go"
 	"github.com/opentracing/opentracing-go/harness"
+	"github.com/stretchr/testify/suite"
 
 	"go.elastic.co/apm"
 	"go.elastic.co/apm/module/apmhttp"
@@ -41,11 +59,42 @@ func TestHarness(t *testing.T) {
 		tracer := New(WithTracer(apmtracer))
 		return tracer, apmtracer.Close
 	}
+
+	var done bool
+	defer func() {
+		if done {
+			recover()
+		}
+	}()
 	harness.RunAPIChecks(t, newTracer,
 		harness.CheckExtract(true),
 		harness.CheckInject(true),
 		harness.UseProbe(harnessAPIProbe{}),
+		func(s *harness.APICheckSuite) {
+			suite.Run(t, harnessSuiteWrapper{s})
+			done = true
+			panic("done") // prevent suite.Run(t, s)
+		},
 	)
+}
+
+type harnessSuiteWrapper struct {
+	*harness.APICheckSuite
+}
+
+func (w harnessSuiteWrapper) TestStartSpanWithParent() {
+	// APICheckSuite.TestStartSpanWithParent tests both child-of and
+	// follows-from. We don't support follows-from, but we don't want
+	// that to prevent us from testing the child-of case.
+	isValidSpanRef = func(ref opentracing.SpanReference) bool {
+		switch ref.Type {
+		case opentracing.ChildOfRef, opentracing.FollowsFromRef:
+			return true
+		}
+		return false
+	}
+	defer func() { isValidSpanRef = isChildOfSpanRef }()
+	w.APICheckSuite.TestStartSpanWithParent()
 }
 
 type harnessAPIProbe struct{}