+++ module/apmot/wrapper.go
@@ -1,9 +1,27 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
 package apmot
 
 import (
 	"context"
 
 	opentracing "github.com/opentracing/opentracing-go"
+	"github.com/opentracing/opentracing-go/log"
 
 	"go.elastic.co/apm"
 	"go.elastic.co/apm/internal/apmcontext"
@@ -111,6 +129,11 @@ func (s apmSpanWrapper) Context() opentracing.SpanContext {
 	return s.spanContext
 }
 
+// BaggageItem returns the empty string; we do not support baggage.
+func (apmSpanWrapper) BaggageItem(key string) string {
+	return ""
+}
+
 // SetBaggageItem is a no-op; we do not support baggage.
 func (s apmSpanWrapper) SetBaggageItem(key, val string) opentracing.Span {
 	// We do not support baggage.
@@ -162,15 +185,18 @@ func (s apmTransactionWrapper) Context() opentracing.SpanContext {
 	return s.spanContext
 }
 
+// BaggageItem returns the empty string; we do not support baggage.
+func (apmTransactionWrapper) BaggageItem(key string) string {
+	return ""
+}
+
 // SetBaggageItem is a no-op; we do not support baggage.
 func (s apmTransactionWrapper) SetBaggageItem(key, val string) opentracing.Span {
 	// We do not support baggage.
 	return s
 }
 
-type apmBaseWrapper struct {
-	unsupportedSpanMethods
-}
+type apmBaseWrapper struct{}
 
 // Tracer returns the Tracer that created this span.
 func (apmBaseWrapper) Tracer() opentracing.Tracer {
@@ -185,3 +211,28 @@ func (apmBaseWrapper) Finish() {}
 // FinishWithOptions is like Finish, but provides explicit control over the
 // end timestamp and log data.
 func (apmBaseWrapper) FinishWithOptions(opentracing.FinishOptions) {}
+
+// LogKV is a no-op for APM wrapper spans.
+func (apmBaseWrapper) LogKV(keyValues ...interface{}) {
+	// No-op.
+}
+
+// LogFields is a no-op for APM wrapper spans.
+func (apmBaseWrapper) LogFields(fields ...log.Field) {
+	// No-op.
+}
+
+// LogEvent is deprecated, and is a no-op.
+func (apmBaseWrapper) LogEvent(event string) {
+	// Deprecated, no-op.
+}
+
+// LogEventWithPayload is deprecated, and is a no-op.
+func (apmBaseWrapper) LogEventWithPayload(event string, payload interface{}) {
+	// Deprecated, no-op.
+}
+
+// Log is deprecated, and is a no-op.
+func (apmBaseWrapper) Log(ld opentracing.LogData) {
+	// Deprecated, no-op.
+}